session cookie implemented

+++ src/app/Products/product-list/product-list.component.ts
@@ -4,6 +4,7 @@ import { ActivatedRoute } from '@angular/router';
 import { Router } from '@angular/router';
 import { ProductService } from '../../../assets/Services/product.service';
 import { constants } from '../../../assets/constant/constant';
+import { AuthService } from '../../../assets/Services/auth.service';
 
 @Component({
   selector: 'app-product-list',
@@ -13,14 +14,24 @@ import { constants } from '../../../assets/constant/constant';
 export class ProductListComponent {
   products: any[];
 
-  constructor( private toastrService: ToastrService , private route: ActivatedRoute , private router: Router , private productService: ProductService ) {
+  constructor( 
+    private toastrService: ToastrService , 
+    private route: ActivatedRoute , 
+    private router: Router , 
+    private productService: ProductService ,
+    private authService: AuthService
+  ) {
     this.products = [];
   }
 
   getProducts() {
-    this.productService.getProducts().subscribe((products: any) => {
-        this.products = products;
-    })
+    
+    if( this.authService.hasValidToken() ) {
+      
+      this.productService.getProducts().subscribe((products: any) => {
+          this.products = products;
+      })
+    }
   }
 
   ngOnInit(): void {
@@ -28,13 +39,17 @@ export class ProductListComponent {
   }
 
   deleteProduct( id: number ) {
-    this.productService.deleteProduct(id).subscribe((success) => {
-      this.toastrService.success(constants.toastrRemoveProductMsg);
-      console.log(success);
-      this.getProducts();
-    },(error) => {
-      console.log(error);
-    })
+    
+    if( this.authService.hasValidToken() ) {
+      
+      this.productService.deleteProduct(id).subscribe((success) => {
+        this.toastrService.success(constants.toastrRemoveProductMsg);
+        console.log(success);
+        this.getProducts();
+      },(error) => {
+        console.log(error);
+      })
+    }
   }
 
   editProduct( prodId: number ) {
